pyproject.toml
[tool.poetry]
name = "django-boilerplate"
version = "0.1.0"
description = ""
authors = ["savilard <vitaliyantonoff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

Django = "3.2.14"
django-split-settings = "^1.1.0"
dj-database-url = "^0.5.0"
djangorestframework = "^3.13.1"
django-behaviors = "^0.5.1"
djoser = "^2.1.0"

psycopg2-binary = "^2.9.3"
environs = "^9.5.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
flake8-django = "^1.1.2"
flake8-aaa = "^0.12.2"

mypy = "^0.961"
black = "^22.6.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
